refactor: notes form mentor(comments, consoles, .env, index, etc)
