+++ .gitignore
@@ -26,3 +26,6 @@ dist-ssr
 # TypeScript build info
 tsconfig.tsbuildinfo
 coverage/
+
+# Ignore environment variables
+.env
