+++ src/components/Search/Search.tsx
@@ -18,7 +18,6 @@ function Search(props: Props) {
     setQueryString(event.target.value);
   }
 
-  console.log('рендер Search.');
   return (
     <>
       <section className={styles.top}>
