+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export { default as ErrorBoundary } from './ErrorBoundary/ErrorBoundary';
+export { default as Header } from './Header/Header';
+export { default as Loader } from './Loader/Loader';
+export { default as Search } from './Search/Search';
+export { default as ResultPage } from './ResultPage/ResultPage';
+export { default as Details } from './Details/Details';
