+++ src/store/slices/apiSlice.ts
@@ -1,8 +1,10 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 
+const BASE_URL = import.meta.env.VITE_API_BASE_URL;
+
 export const apiSlice = createApi({
   reducerPath: 'api',
-  baseQuery: fetchBaseQuery({ baseUrl: 'https://swapi.dev/api/' }),
+  baseQuery: fetchBaseQuery({ baseUrl: BASE_URL || 'https://swapi.dev/api/' }),
   endpoints: (builder) => ({
     getCharacters: builder.query({
       query: ({ query, page }) =>
